:arrow_up: Update @staart/api to v3 beta

+++ src/crons/daily.ts
@@ -1,7 +1,7 @@
-import { elasticSearch } from "@staart/elasticsearch";
+import { elasticSearch, elasticSearchEnabled } from "@staart/elasticsearch";
 import { ms } from "@staart/text";
 import { CronJob } from "cron";
-import { ELASTIC_LOGS_PREFIX } from "../config";
+import { ELASTIC_LOGS_INDEX } from "../config";
 
 export default () => {
   new CronJob(
@@ -15,24 +15,25 @@ export default () => {
 };
 
 const deleteOldLogs = async () => {
-  return (
-    await elasticSearch.deleteByQuery({
-      index: `${ELASTIC_LOGS_PREFIX}*`,
-      body: {
-        query: {
-          bool: {
-            must: [
-              {
-                range: {
-                  date: {
-                    lte: new Date(new Date().getTime() - ms("92 days"))
-                  }
-                }
-              }
-            ]
-          }
-        }
-      }
-    })
-  ).body;
+  if (elasticSearchEnabled)
+    return (
+      await elasticSearch.deleteByQuery({
+        index: ELASTIC_LOGS_INDEX,
+        body: {
+          query: {
+            bool: {
+              must: [
+                {
+                  range: {
+                    date: {
+                      lte: new Date(new Date().getTime() - ms("92 days")),
+                    },
+                  },
+                },
+              ],
+            },
+          },
+        },
+      })
+    ).body;
 };
